feat: add PersistenceError to decouple from sqlx error
